+++ resources/lang/en/auth.php
@@ -15,5 +15,46 @@ return [
 
     'failed' => 'These credentials do not match our records.',
     'throttle' => 'Too many login attempts. Please try again in :seconds seconds.',
-
+    'title' => [
+        'login' => 'Login',
+        'register' => 'Register',
+        'forgot_password' => 'Forgot Your Password?',
+        'reset_password' => 'Reset Password',
+    ],
+    'label' => [
+        'email' => 'Email Address',
+        'password' => 'Password',
+        'remember' => 'Remember Me'
+    ],
+    'placeholder' => [
+        'name' => 'Full Name',
+        'email' => 'Email Address',
+        'password' => 'Password',
+        'new_password' => 'New Password',
+        'password_confirmation' => 'Password Confirmation',
+    ],
+    'icon' => [
+        'name' => '<i class="fa fa-user form-control-feedback"></i>',
+        'email' => '<i class="fa fa-envelope form-control-feedback"></i>',
+        'password' => '<i class="fa fa-lock form-control-feedback"></i>',
+        'password_confirmation' => '<i class="fa fa-check form-control-feedback"></i>',
+    ],
+    'button' => [
+        'sign_in'  => 'Sign In',
+        'sign_out' => 'Sign Out',
+        'register' => 'Register',
+        'forgot_password' => 'Send Password Reset Link',
+        'reset_password' => 'Reset Password',
+    ],
+    'text' => [
+        'forgot_password' => 'Forgot Your Password?',
+        'register' => 'Register a new membership',
+        'login' => 'I already have a membership',
+    ],
+    'message' => [
+        'login' => 'Sign in to start your session',
+        'register' => 'Fill all fields for register our system',
+        'forgot_password' => 'Please enter your email address for reset password link.',
+        'reset_password' => 'Place enter your email address and new password',
+    ]
 ];