+++ resources/views/auth/login.blade.php
@@ -1,71 +1,37 @@
-@extends('layouts.app')
+@extends('templates.adminlte.layouts.auth')
+
+@section('title', trans('auth.title.login'))
 
 @section('content')
-<div class="container">
-    <div class="row justify-content-center">
-        <div class="col-md-8">
-            <div class="card">
-                <div class="card-header">{{ __('Login') }}</div>
+    <div class="login-box-body">
+        @include('templates.adminlte.partials.alerts')
+        <p class="login-box-msg">{{ trans('auth.message.login') }}</p>
 
-                <div class="card-body">
-                    <form method="POST" action="{{ route('login') }}" aria-label="{{ __('Login') }}">
-                        @csrf
-
-                        <div class="form-group row">
-                            <label for="email" class="col-sm-4 col-form-label text-md-right">{{ __('E-Mail Address') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="email" type="email" class="form-control{{ $errors->has('email') ? ' is-invalid' : '' }}" name="email" value="{{ old('email') }}" required autofocus>
-
-                                @if ($errors->has('email'))
-                                    <span class="invalid-feedback" role="alert">
-                                        <strong>{{ $errors->first('email') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
-
-                        <div class="form-group row">
-                            <label for="password" class="col-md-4 col-form-label text-md-right">{{ __('Password') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="password" type="password" class="form-control{{ $errors->has('password') ? ' is-invalid' : '' }}" name="password" required>
-
-                                @if ($errors->has('password'))
-                                    <span class="invalid-feedback" role="alert">
-                                        <strong>{{ $errors->first('password') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
-
-                        <div class="form-group row">
-                            <div class="col-md-6 offset-md-4">
-                                <div class="form-check">
-                                    <input class="form-check-input" type="checkbox" name="remember" id="remember" {{ old('remember') ? 'checked' : '' }}>
-
-                                    <label class="form-check-label" for="remember">
-                                        {{ __('Remember Me') }}
-                                    </label>
-                                </div>
-                            </div>
-                        </div>
-
-                        <div class="form-group row mb-0">
-                            <div class="col-md-8 offset-md-4">
-                                <button type="submit" class="btn btn-primary">
-                                    {{ __('Login') }}
-                                </button>
-
-                                <a class="btn btn-link" href="{{ route('password.request') }}">
-                                    {{ __('Forgot Your Password?') }}
-                                </a>
-                            </div>
-                        </div>
-                    </form>
+        {!! Form::open(['route' => 'login', 'method' => 'POST']) !!}
+            <div class="form-group has-feedback">
+                {!! Form::email('email', null, ['class' => 'form-control' , 'placeholder' => trans('auth.placeholder.email')]) !!}
+                {!! trans('auth.icon.email') !!}
+            </div>
+            <div class="form-group has-feedback">
+                {!! Form::password('password', ['class' => 'form-control' , 'placeholder' => trans('auth.placeholder.password')]) !!}
+                {!! trans('auth.icon.password') !!}
+            </div>
+            <div class="row">
+                <div class="col-xs-8">
+                    <div class="checkbox">
+                        <label>{!! Form::checkbox('remember', 1, old('remember')) !!} {{ trans('auth.label.remember') }}</label>
+                    </div>
+                </div>
+                <div class="col-xs-4">
+                    {!! Form::button(trans('auth.button.sign_in'), ['type' => 'submit', 'class' => 'btn btn-block btn-flat btn-primary']) !!}
                 </div>
             </div>
+        {!! Form::close() !!}
+        <div class="text-center">
+            <p>- {{ __('OR') }} -</p>
+            <a class="btn btn-link" href="{{ route('password.request') }}">{{ trans('auth.text.forgot_password') }}</a><br>
+            <a class="btn btn-link" href="{{ route('register') }}">{{ trans('auth.text.register') }}</a>
         </div>
     </div>
-</div>
+    <!-- /.login-box-body -->
 @endsection