+++ resources/views/auth/register.blade.php
@@ -1,77 +1,37 @@
-@extends('layouts.app')
+@extends('templates.adminlte.layouts.auth')
+
+@section('title', trans('auth.title.register'))
 
 @section('content')
-<div class="container">
-    <div class="row justify-content-center">
-        <div class="col-md-8">
-            <div class="card">
-                <div class="card-header">{{ __('Register') }}</div>
+    <div class="login-box-body">
+        @include('templates.adminlte.partials.alerts')
+        <p class="login-box-msg">{{ trans('auth.message.register') }}</p>
 
-                <div class="card-body">
-                    <form method="POST" action="{{ route('register') }}" aria-label="{{ __('Register') }}">
-                        @csrf
-
-                        <div class="form-group row">
-                            <label for="name" class="col-md-4 col-form-label text-md-right">{{ __('Name') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="name" type="text" class="form-control{{ $errors->has('name') ? ' is-invalid' : '' }}" name="name" value="{{ old('name') }}" required autofocus>
-
-                                @if ($errors->has('name'))
-                                    <span class="invalid-feedback" role="alert">
-                                        <strong>{{ $errors->first('name') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
-
-                        <div class="form-group row">
-                            <label for="email" class="col-md-4 col-form-label text-md-right">{{ __('E-Mail Address') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="email" type="email" class="form-control{{ $errors->has('email') ? ' is-invalid' : '' }}" name="email" value="{{ old('email') }}" required>
-
-                                @if ($errors->has('email'))
-                                    <span class="invalid-feedback" role="alert">
-                                        <strong>{{ $errors->first('email') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
-
-                        <div class="form-group row">
-                            <label for="password" class="col-md-4 col-form-label text-md-right">{{ __('Password') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="password" type="password" class="form-control{{ $errors->has('password') ? ' is-invalid' : '' }}" name="password" required>
-
-                                @if ($errors->has('password'))
-                                    <span class="invalid-feedback" role="alert">
-                                        <strong>{{ $errors->first('password') }}</strong>
-                                    </span>
-                                @endif
-                            </div>
-                        </div>
-
-                        <div class="form-group row">
-                            <label for="password-confirm" class="col-md-4 col-form-label text-md-right">{{ __('Confirm Password') }}</label>
-
-                            <div class="col-md-6">
-                                <input id="password-confirm" type="password" class="form-control" name="password_confirmation" required>
-                            </div>
-                        </div>
-
-                        <div class="form-group row mb-0">
-                            <div class="col-md-6 offset-md-4">
-                                <button type="submit" class="btn btn-primary">
-                                    {{ __('Register') }}
-                                </button>
-                            </div>
-                        </div>
-                    </form>
-                </div>
+        {!! Form::open(['route' => 'register', 'method' => 'POST']) !!}
+            <div class="form-group has-feedback">
+                {!! Form::text('name', null, ['class' => 'form-control' , 'placeholder' => trans('auth.placeholder.name')]) !!}
+                {!! trans('auth.icon.name') !!}
             </div>
+            <div class="form-group has-feedback">
+                {!! Form::email('email', null, ['class' => 'form-control' , 'placeholder' => trans('auth.placeholder.email')]) !!}
+                {!! trans('auth.icon.email') !!}
+            </div>
+            <div class="form-group has-feedback">
+                {!! Form::password('password', ['class' => 'form-control' , 'placeholder' => trans('auth.placeholder.password')]) !!}
+                {!! trans('auth.icon.password') !!}
+            </div>
+            <div class="form-group has-feedback">
+                {!! Form::password('password_confirmation', ['class' => 'form-control' , 'placeholder' => trans('auth.placeholder.password_confirmation')]) !!}
+                {!! trans('auth.icon.password_confirmation') !!}
+            </div>
+            <div class="clearfix">
+                {!! Form::button(trans('auth.button.register'), ['type' => 'submit', 'class' => 'btn btn-block btn-flat btn-primary']) !!}
+            </div>
+        {!! Form::close() !!}
+        <div class="text-center">
+            <p>- {{ __('OR') }} -</p>
+            <a class="btn btn-link" href="{{ route('login') }}">{{ trans('auth.text.login') }}</a>
         </div>
     </div>
-</div>
+    <!-- /.login-box-body -->
 @endsection