+++ resources/views/templates/adminlte/partials/alerts.blade.php
@@ -17,3 +17,9 @@
         {{ session('error') }}
     </div>
 @endif
+
+@if (session('status'))
+    <div class="alert alert-info" role="alert">
+        {{ session('status') }}
+    </div>
+@endif